Initial commit / added most up-to-date files
Program is capable of turning the camera at this point, and objects/ray intersections are not yet implemented. It seems as though the camera may not be aligned with the world's axes or the rotations may not be properly aligned with the camera's/world's axes, but this may just be a side effect of the method for obtaining the default color for rays.